malloc replaced(list_t)

+++ include/hash.h
@@ -33,9 +33,9 @@ typedef struct hash_t {
     bucket_t* bucket_list;
 } hash_t;
 
-int hash_get_bucket_count(int max_entry);
+u32 hash_get_bucket_count(u32 max_entry);
 
-int hash_calculate_bucket_size(int max_entry);
+u32 hash_calculate_bucket_size(u32 max_entry);
 
 /**
  * @fn hash_init
@@ -48,7 +48,7 @@ int hash_calculate_bucket_size(int max_entry);
  * @return NULL  - when out of memory.
  * @return pointer to hash table
  */
-void* hash_init(int max_entry, int key_size, LIBCACHE_CMP_KEY* key_cmp, LIBCACHE_KEY_TO_NUMBER* key_to_num, void *pool_handle);
+void* hash_init(u32 max_entry, size_t key_size, LIBCACHE_CMP_KEY* key_cmp, LIBCACHE_KEY_TO_NUMBER* key_to_num, void *pool_handle);
 
 /**
  * @fn hash_add
